requirements.txt
aiogram==2.9.2
Telethon==1.15.0
python-dotenv==0.14.0
